Moved keysFromQuery.ts to utils directory.
